refactor: own lib for EspFlasher
