+++ .licenserc.json
@@ -1,5 +1,5 @@
 {
-  "{include,src}/**/*.{cpp,hpp}": [
+  "{libs,src}/**/*.{cpp,hpp}": [
     "// This Source Code Form is subject to the terms of the Mozilla Public",
     "// License, v. 2.0. If a copy of the MPL was not distributed with this",
     "// file, You can obtain one at https://mozilla.org/MPL/2.0/."
