+++ libs/esp_flasher/CMakeLists.txt
@@ -1,0 +1,43 @@
+file(GLOB_RECURSE SRC include/*.hpp src/*.cpp)
+add_library(ESPFlasher ${SRC})
+add_library(Qt::ESPFlasher ALIAS ESPFlasher)
+
+target_compile_features(ESPFlasher PUBLIC cxx_std_23)
+
+target_include_directories(
+  ESPFlasher SYSTEM
+  INTERFACE include
+  PRIVATE include/esp_flasher)
+
+target_common_warnings(ESPFlasher PRIVATE)
+
+if(NOT TARGET flasher)
+  cpmaddpackage(
+    NAME
+    esp-serial-flasher
+    GITHUB_REPOSITORY
+    "espressif/esp-serial-flasher"
+    VERSION
+    1.4.1
+    SYSTEM
+    ON
+    OPTIONS
+    "PORT USER_DEFINED"
+    # "MD5_ENABLED OFF" # Currently not working?!
+    "SERIAL_FLASHER_STUB_PULL_VERSION 1.2.0"
+    "SERIAL_FLASHER_STUB_PULL_SOURCE https://github.com/espressif/esptool-legacy-flasher-stub/releases/download"
+  )
+endif()
+
+if(NOT TARGET magic_enum)
+  cpmaddpackage("gh:Neargye/magic_enum@0.9.6")
+endif()
+
+if(NOT TARGET Microsoft.GSL::GSL)
+  cpmaddpackage("gh:microsoft/GSL@4.1.0")
+endif()
+
+target_link_libraries(
+  ESPFlasher
+  PUBLIC flasher Qt::Core Qt::SerialPort Qt::Widgets
+  PRIVATE magic_enum::magic_enum Microsoft.GSL::GSL)
