+++ libs/esp_flasher/include/esp_flasher/bin.hpp
