+++ libs/esp_flasher/src/available_ports.cpp
